// part9/patientor/backend/src/types.ts
export enum Gender {
    Male = "male",
    Female = "female",
    Other = "other"
}

export interface Discharge {
    date: string;
    criteria: string;
}

export interface SickLeave {
    startDate: string;
    endDate: string;
}

export interface EntryOptions {
    value: EntryType;
    label: string;
}

export interface PatientEntry {
    id: string;
    name: string;
    dateOfBirth: string;
    ssn?: string;
    gender: Gender;
    occupation: string;
    entries: Entry[];
}

export interface DiagnosisEntry {
    code: string;
    name: string;
    latin?: string;
}

export interface BaseEntry {
    id: string;
    description: string;
    date: string;
    specialist: string;
    type: EntryType;
    diagnosisCodes?: Array<DiagnosisEntry['code']>;
}

export type NonSensitiveBaseEntry = Omit<BaseEntry, 'id'>

export enum HealthCheckRating {
    "Healthy" = 0,
    "LowRisk" = 1,
    "HighRisk" = 2,
    "CriticalRisk" = 3
}

export enum EntryType {
    HospitalEntry = "Hospital",
    OccupationalHealthcareEntry = "OccupationalHealthcare",
    HealthCheckEntry = "HealthCheck",
}  
  
export interface HealthCheckEntry extends BaseEntry {
    type: EntryType.HealthCheckEntry;
    healthCheckRating: HealthCheckRating;
}

export interface HospitalEntry extends BaseEntry {
    type: EntryType.HospitalEntry;
    discharge: Discharge;
}

export interface OccupationalHealthcareEntry extends BaseEntry {
    type: EntryType.OccupationalHealthcareEntry;
    employerName: string;
    sickLeave?: SickLeave;
}

export type Entry =
    | HospitalEntry
    | OccupationalHealthcareEntry
    | HealthCheckEntry;

// Define special omit for unions
export type UnionOmit<T, K extends string | number | symbol> = T extends unknown ? Omit<T, K> : never;
// Define Entry without the 'id' property
export type EntryWithoutId = UnionOmit<Entry, 'id'>;